# Update storage weights