--- a/runtime/polkadot/constants/src/weights/paritydb_weights.rs
+++ b/runtime/polkadot/constants/src/weights/paritydb_weights.rs
@@ -15,6 +15,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 4.0.0-dev
+//! DATE: 2022-03-30 (Y/M/D)
+//!
+//! DATABASE: `ParityDb`, RUNTIME: `Polkadot`
+//! BLOCK-NUM: `BlockId::Number(9653477)`
+//! SKIP-WRITE: `false`, SKIP-READ: `false`, WARMUPS: `1`
+//! STATE-VERSION: `V0`, STATE-CACHE-SIZE: `0`
+//! WEIGHT-PATH: `runtime/polkadot/constants/src/weights/`
+//! METRIC: `Average`, WEIGHT-MUL: `1.1`, WEIGHT-ADD: `0`
+
+// Executed Command:
+//   ./target/production/polkadot
+//   benchmark-storage
+//   --db=paritydb
+//   --state-version=0
+//   --mul=1.1
+//   --weight-path=runtime/polkadot/constants/src/weights/
+
+/// Storage DB weights for the `Polkadot` runtime and `ParityDb`.
 pub mod constants {
 	use frame_support::{
 		parameter_types,
@@ -25,8 +44,35 @@ pub mod constants {
 		/// `ParityDB` can be enabled with a feature flag, but is still experimental. These weights
 		/// are available for brave runtime engineers who may want to try this out as default.
 		pub const ParityDbWeight: RuntimeDbWeight = RuntimeDbWeight {
-			read: 8_000 * constants::WEIGHT_PER_NANOS,
-			write: 50_000 * constants::WEIGHT_PER_NANOS,
+			/// Time to read one storage item.
+			/// Calculated by multiplying the *Average* of all values with `1.1` and adding `0`.
+			///
+			/// Stats [NS]:
+			///   Min, Max: 4_611, 13_478_005
+			///   Average:  10_750
+			///   Median:   10_655
+			///   Std-Dev:  12214.49
+			///
+			/// Percentiles [NS]:
+			///   99th: 14_451
+			///   95th: 12_588
+			///   75th: 11_200
+			read: 11_826 * constants::WEIGHT_PER_NANOS,
+
+			/// Time to write one storage item.
+			/// Calculated by multiplying the *Average* of all values with `1.1` and adding `0`.
+			///
+			/// Stats [NS]:
+			///   Min, Max: 8_023, 47_367_740
+			///   Average:  34_592
+			///   Median:   32_703
+			///   Std-Dev:  49417.24
+			///
+			/// Percentiles [NS]:
+			///   99th: 69_379
+			///   95th: 47_168
+			///   75th: 35_252
+			write: 38_052 * constants::WEIGHT_PER_NANOS,
 		};
 	}
 
@@ -39,7 +85,7 @@ pub mod constants {
 		// NOTE: If this test fails but you are sure that the generated values are fine,
 		// you can delete it.
 		#[test]
-		fn sane() {
+		fn bound() {
 			// At least 1 µs.
 			assert!(
 				W::get().reads(1) >= constants::WEIGHT_PER_MICROS,