Update RocksDB storage weights.  
- Read: 25µs -> 20.5µs
- Write: 100µs -> 83.47µs

Update ParityDB storage weigts:
- Read: 8µs -> 11.8µs
- Write: 50µs -> 38µs

We add 10% to the average to account for future state growth.  
I could do the same for Kusama if needed.  

How to reproduce these results on reference hardware:  
1. Optional: Download a snapshot from [Polkachu](https://polkachu.com/snapshots/) if you don't have a synced node
2. Run the benchmark and generate the weight files (add a `--db=paritydb` for ParityDb).
You may need to pass in the `--base-path` command if you extracted the snapshot into a different directory.
```sh
./target/production/polkadot benchmark-storage --state-version=0 --mul=1.1 --weight-path=runtime/polkadot/constants/src/weights/
```